+++ backend/src/app/interactor/patientInteractor.ts
@@ -26,21 +26,10 @@ export default class PatientInteractor {
 
   async getAll(): Promise<Patient[]> {
     const [patients, error] = await wrapError(this.patientRepository.findAll());
-    const patientsActive: Patient[] = [];
-
-    /* patients.forEach(async element => {
-      const record : Record = await this.getRecord(element.recordId);
-      if (record.active === true) {
-        console.log("activo");
-        patientsActive.push(element);
-      }
-    }); */
 
     if (error) {
       throw error;
     }
-
-    console.log(patientsActive);
 
     return this.patientPresenter.findAll(patients);
   }
