+++ frontend/src/api/forms.ts
@@ -3,40 +3,40 @@ import server from 'src/utils/server';
 
 export async function registerForm(form: Form) {
   const result = await server.post('/forms', form);
-  return result;
+  return result.data || result;
 }
 
 export async function registerPatientForm(id: number, form: Form) {
   const result = await server.post(`/records/${id}/patientForms`, form);
-  return result;
+  return result.data || result;
 }
 
 export async function updatePatientForm(id: number, form: Form) {
   const result = await server.put(`/patientFormField/${id}`, form);
-  return result;
+  return result.data || result;
 }
 
 export async function listFormsWithRecordId(id: number) {
   const result = await server.get(`/records/${id}/forms`);
-  return result;
+  return result.data || result;
 }
 
 export async function getFormField(id: number) {
   const result = await server.get(`/patientFormField/${id}`);
-  return result;
+  return result.data || result;
 }
 
 export async function getForms() {
   const result = await server.get('/forms');
-  return result;
+  return result.data || result;
 }
 
 export async function deleteForm(id: number) {
   const result = await server.delete(`/forms/${id}`);
-  return result;
+  return result.data || result;
 }
 
 export async function getForm(id: number) {
   const result = await server.get(`/forms/${id}`);
-  return result.data;
+  return result.data || result;
 }
