+++ frontend/src/api/patient.tsx
@@ -1,34 +1,32 @@
-import Record from 'src/interfaces/record';
 import Patient from '../interfaces/patient';
 import server from '../utils/server';
-import handleResponse from '../utils/handleResponse';
 
 export async function createPatient(patient:Patient) {
-  const result = await server.post('/patients', { ...patient }).then(handleResponse).catch(handleResponse);
-  return result.data;
+  const result = await server.post('/patients', { ...patient });
+  return result.data || result;
 }
 
 export async function createCouple(patient:Array<Patient>) {
-  const result = await server.post('/patients', patient).then(handleResponse).catch(handleResponse);
-  return result.data;
+  const result = await server.post('/patients', patient);
+  return result.data || result;
 }
 
 export async function getPatients() {
-  const result = await server.get('/patients').then(handleResponse).catch(handleResponse);
-  return result;
+  const result = await server.get('/patients');
+  return result.data || result;
 }
 
 export async function getPatientRecord(id: number) {
-  const result = await server.get<Record>(`records/${id}`);
-  return result;
+  const result = await server.get(`records/${id}`);
+  return result.data || result;
 }
 
 export async function canalizePatient(patient:Array<Patient>) {
-  const result = await server.post('/canalize-patient', patient).then(handleResponse).catch(handleResponse);
-  return result.data;
+  const result = await server.post('/canalize-patient', patient);
+  return result.data || result;
 }
 
 export async function archiveRecord(id: number) {
-  const result = await server.put(`/archive-record/${id}`).then(handleResponse).catch(handleResponse);
-  return result.data;
+  const result = await server.put(`/archive-record/${id}`);
+  return result.data || result;
 }
