+++ frontend/src/components/loadingSpinner.tsx
@@ -1,4 +1,4 @@
-import { CircularProgress, makeStyles } from '@material-ui/core';
+import { CircularProgress, Container, makeStyles } from '@material-ui/core';
 import React from 'react';
 
 const useStyles = makeStyles(() => ({
@@ -7,15 +7,18 @@ const useStyles = makeStyles(() => ({
     justifyContent: 'center',
     alignItems: 'center',
     height: '100%',
+    flexDirection: 'column',
+    minHeight: '80vh',
+    flex: 1,
   },
 }));
 function LoadingSpinner() {
   const classes = useStyles();
 
   return (
-    <div className={classes.loading}>
-      <CircularProgress />
-    </div>
+    <Container className={classes.loading}>
+      <CircularProgress size={100} />
+    </Container>
   );
 }
