+++ frontend/src/utils/promiseLoader.tsx
@@ -3,9 +3,9 @@ import React, { useEffect, useState } from 'react';
 import LoadingSpinner from 'src/components/loadingSpinner';
 
 function PromiseLoader<T>(
-  promise: Promise<AxiosResponse<any>>,
+  promise: () => Promise<AxiosResponse<any>>,
   onLoad: (data: T) => JSX.Element,
-  onError: (error: AxiosError) => JSX.Element,
+  onError?: (error: AxiosError) => JSX.Element,
 ) {
   const [data, setData] = useState<T>();
   const [error, setError] = useState<AxiosError>();
@@ -13,8 +13,9 @@ function PromiseLoader<T>(
   useEffect(() => {
     const fetchData = async () => {
       try {
-        const response = await promise;
-        setData(response.data);
+        const response = await promise();
+        const result = response.data ? response.data : response;
+        setData(result);
       } catch (e) {
         const err = e as AxiosError;
         setError(err);
@@ -24,8 +25,16 @@ function PromiseLoader<T>(
   }, []);
 
   if (error) {
-    console.error(error);
-    return onError(error);
+    if (onError) {
+      return onError(error);
+    }
+    // default error handling
+    switch (error.response?.status) {
+      case 404:
+        return <h2>No se encontró el expediente</h2>;
+      default:
+        return <h2>Ocurrió un error de conexión.</h2>;
+    }
   }
   if (data === undefined) {
     console.log('Loading content...');
