+++ frontend/src/views/Forms/newPatientForm.tsx
@@ -1,26 +1,19 @@
 import React from 'react';
 
-import server from 'src/utils/server';
 import 'src/App.css';
 import { useParams } from 'react-router-dom';
 import PromiseLoader from 'src/utils/promiseLoader';
 import Form from 'src/interfaces/form';
 import GenerateForm from 'src/components/Forms/NewPatientForm/generateForm';
+import { getForm } from 'src/api/forms';
 
 function NewPatientForm() {
   const { id, formId } : any = useParams();
-  const mPromise = server.get<Form>(`/forms/${formId}`);
+  const mPromise = getForm;
   const content = PromiseLoader<Form>(
-    mPromise,
+    () => mPromise(formId),
     (form) => <GenerateForm data={form} id={id} />,
-    (error) => {
-      switch (error.response?.status) {
-        case 404:
-          return <h2>No se encontró el expediente</h2>;
-        default:
-          return <h2>Ocurrió un error de conexión.</h2>;
-      }
-    },
+
   );
   return content;
 }
